#172 - Spanish translations for landing page (#173)
This is the initial version of the Spanish translations. Since I'm new
and also new in Boston, I'm missing a lot of the context necessary for
the translations to be as accurate as possible. So I'm going to add
comments in my own code to open a discussion about the translations in
which I need more information or to ask for advice as to what seems
better to others. I'll attempt to provide enough context that even if
you don't speak Spanish my questions make sense.