Add automated linter commenting to PRs (#307)
This diff changes three things, the first of which is important. The
latter two are incidental.
- **Introduces a github action that runs on PR creations and updates.
This job will comment linting errors on the PR, removing comments as
linting errors are resolved in subsequent commits. This job does not
block a PR from merging.**
- **Lint Fix**: Moves buttonVariants declaration to a separate file. The
linter warned that exporting them in the same file interferes with react
fast-refresh
- **Lint Fix**: Removes _ as an argument in a locale fallback function.
This resolves a "no unused variables" linter warning. We can later
discuss whether `[, value]` or `[_, value]` is better syntax. We will
have to fight with the linter a bit to enable the latter.

We have an existing eslint config from project bootstrapping. It
includes recommended defaults, and plugins for react-hooks and
react-refresh. Other than that it is fairly minimal.

Conversations about code style should precede changes to that config. We
have already had discussion around alerting contributors to our
standards around color. This would be a good immediate change to that
config.

An example of this process in action can be seen in [this PR on my
fork](https://github.com/MattClarke131/boston-liquor-license-tracker/pull/13).