Fix/data correction (#311)
**Manual Corrections** 
- Applicants from the 05/29/25 voting minutes were not captured during
the initial historical data backfill.
- Milkweed was included in the 05/29/25 voting minutes and therefore was
not present in the dataset when the February 13 results were reviewed.
Since application record was missing, the status could not be manually
updated at that time. Once the missing application was identified and
added, the status was manually set to “Granted” to reflect the official
outcome.
- DTDA appeared in the February 13 results, but its status was missed
during the manual review. Because status updates are currently handled
entirely manually, this was a human oversight, and the status has since
been corrected to “Granted.”